# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lume - AI Assistant Platform",
    "description": "Platform for creating and managing AI assistants with RAG",
    "version": "2.0.0"
  },
  "paths": {
...
    "/evaluation/datasets": {
      "get": {
        "tags": [
          "evaluation"
        ],
        "summary": "List Datasets",
        "description": "Retrieve all evaluation datasets.",
        "operationId": "listDatasets",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "evaluation"
        ],
        "summary": "Create Dataset",
        "description": "Create a dataset with manually provided QA pairs.",
        "operationId": "createDataset",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/evaluation/datasets/{dataset_id}": {
      "put": {
        "tags": [
          "evaluation"
        ],
        "summary": "Update Dataset",
        "description": "Update an evaluation dataset.",
        "operationId": "updateDataset",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "evaluation"
        ],
        "summary": "Delete Dataset",
        "description": "Delete an evaluation dataset.",
        "operationId": "deleteDataset",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/evaluation/ragas": {
      "post": {
        "tags": [
          "evaluation"
        ],
        "summary": "Generate Ragas Dataset",
        "description": "Generate a synthetic QA dataset using RAGAS.",
        "operationId": "generateRagasDataset",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RagasGenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/evaluation/evaluate-assistant": {
      "post": {
        "tags": [
          "evaluation"
        ],
        "summary": "Evaluate Assistant",
        "description": "Evaluate an assistant's performance using RAGAS metrics.",
        "operationId": "evaluateAssistant",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateAssistantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/evaluation/evaluations/{evaluation_id}": {
      "get": {
        "tags": [
          "evaluation"
        ],
        "summary": "Get Evaluation",
        "description": "Retrieve a specific evaluation by ID.",
        "operationId": "getEvaluation",
        "parameters": [
          {
            "name": "evaluation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Evaluation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/evaluation/datasets/{dataset_name}/evaluations": {
      "get": {
        "tags": [
          "evaluation"
        ],
        "summary": "List Evaluations By Dataset",
        "description": "Get all evaluations for a specific dataset.",
        "operationId": "listEvaluationsByDataset",
        "parameters": [
          {
            "name": "dataset_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DatasetCreateRequest": {
        "properties": {
          "dataset_name": {
            "type": "string",
            "title": "Dataset Name",
            "description": "Name of the dataset"
          },
          "qa_pairs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Qa Pairs",
            "description": "List of question-answer pair dicts"
          }
        },
        "type": "object",
        "required": [
          "dataset_name",
          "qa_pairs"
        ],
        "title": "DatasetCreateRequest"
      },
      "DatasetListResponse": {
        "properties": {
          "datasets": {
            "items": {
              "$ref": "#/components/schemas/DatasetResponse"
            },
            "type": "array",
            "title": "Datasets"
          }
        },
        "type": "object",
        "required": [
          "datasets"
        ],
        "title": "DatasetListResponse"
      },
      "DatasetResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "source_collection": {
            "type": "string",
            "title": "Source Collection"
          },
          "generated_at": {
            "type": "string",
            "title": "Generated At"
          },
          "generator": {
            "type": "string",
            "title": "Generator"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "num_pairs": {
            "type": "integer",
            "title": "Num Pairs"
          },
          "qa_pairs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Qa Pairs"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "source_collection",
          "generated_at",
          "generator",
          "model",
          "num_pairs",
          "qa_pairs"
        ],
        "title": "DatasetResponse"
      },
      "DatasetUpdateRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "qa_pairs": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qa Pairs"
          }
        },
        "type": "object",
        "title": "DatasetUpdateRequest",
        "description": "Partial update for a dataset. All fields optional."
      },
      "EvaluateAssistantRequest": {
        "properties": {
          "dataset_name": {
            "type": "string",
            "title": "Dataset Name",
            "description": "Name of the evaluation dataset"
          },
          "assistant_id": {
            "type": "string",
            "title": "Assistant Id",
            "description": "ID of the assistant being evaluated"
          },
          "questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Questions"
          },
          "ground_truths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ground Truths"
          },
          "answers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Answers"
          },
          "retrieved_contexts": {
            "items": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Retrieved Contexts",
            "description": "Retrieved context chunks per question"
          },
          "eval_llm_model": {
            "type": "string",
            "title": "Eval Llm Model",
            "default": "gpt-4o-mini"
          },
          "eval_llm_provider": {
            "type": "string",
            "title": "Eval Llm Provider",
            "default": "openai"
          }
        },
        "type": "object",
        "required": [
          "dataset_name",
          "assistant_id",
          "questions",
          "ground_truths",
          "answers",
          "retrieved_contexts"
        ],
        "title": "EvaluateAssistantRequest"
      },
      "EvaluationListResponse": {
        "properties": {
          "evaluations": {
            "items": {
              "$ref": "#/components/schemas/EvaluationResponse"
            },
            "type": "array",
            "title": "Evaluations"
          }
        },
        "type": "object",
        "required": [
          "evaluations"
        ],
        "title": "EvaluationListResponse"
      },
      "EvaluationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "dataset_name": {
            "type": "string",
            "title": "Dataset Name"
          },
          "assistant_id": {
            "type": "string",
            "title": "Assistant Id"
          },
          "eval_llm_model": {
            "type": "string",
            "title": "Eval Llm Model"
          },
          "eval_llm_provider": {
            "type": "string",
            "title": "Eval Llm Provider"
          },
          "metrics": {
            "$ref": "#/components/schemas/MetricsSummary"
          },
          "detailed_results": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Detailed Results"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "dataset_name",
          "assistant_id",
          "eval_llm_model",
          "eval_llm_provider",
          "metrics",
          "detailed_results",
          "created_at"
        ],
        "title": "EvaluationResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MetricsSummary": {
        "properties": {
          "faithfulness": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Faithfulness"
          },
          "context_recall": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Recall"
          },
          "answer_relevancy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Relevancy"
          },
          "context_precision": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Precision"
          }
        },
        "type": "object",
        "title": "MetricsSummary"
      },
...
      "RagasGenerationRequest": {
        "properties": {
          "collection_name": {
            "type": "string",
            "title": "Collection Name"
          },
          "dataset_name": {
            "type": "string",
            "title": "Dataset Name"
          },
          "testset_size": {
            "type": "integer",
            "title": "Testset Size",
            "description": "Number of test cases to generate",
            "default": 1
          },
          "model_name": {
            "type": "string",
            "title": "Model Name",
            "description": "LLM model for generation",
            "default": "gpt-4o-mini"
          }
        },
        "type": "object",
        "required": [
          "collection_name",
          "dataset_name"
        ],
        "title": "RagasGenerationRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}